AS7-6015 Fix CDI interceptor ordering as per the spec clarification on CDI 1.1
