+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/weld/ejb/interceptor/ordering/CdiInterceptor.java
@@ -14,35 +14,27 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.as.test.integration.weld.interceptor.ordering;
+package org.jboss.as.test.integration.weld.ejb.interceptor.ordering;
 
 import java.io.Serializable;
 import java.util.List;
 
-import javax.ejb.Stateful;
-import javax.enterprise.context.ApplicationScoped;
 import javax.interceptor.AroundInvoke;
-import javax.interceptor.Interceptors;
+import javax.interceptor.Interceptor;
 import javax.interceptor.InvocationContext;
 
-@Stateful
-@ApplicationScoped
+@Interceptor
 @CdiIntercepted
-@Interceptors(LegacyInterceptor.class)
-public class InterceptedBean implements Serializable {
+public class CdiInterceptor implements Serializable {
 
-    private static final long serialVersionUID = -4444919869290540443L;
-
-    public void ping(List<String> list) {
-        list.add("InterceptedBean");
-    }
+    private static final long serialVersionUID = -5949866804898740300L;
 
     @AroundInvoke
     Object aroundInvoke(InvocationContext ctx) throws Exception {
         Object[] parameters = ctx.getParameters();
         @SuppressWarnings("unchecked")
         List<String> sequence = (List<String>) parameters[0];
-        sequence.add("TargetClassInterceptor");
+        sequence.add("CdiInterceptor");
         return ctx.proceed();
     }
 }
