+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/weld/ejb/interceptor/ordering/InterceptedBean.java
@@ -14,24 +14,35 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.as.test.integration.weld.interceptor.ordering;
+package org.jboss.as.test.integration.weld.ejb.interceptor.ordering;
 
 import java.io.Serializable;
 import java.util.List;
 
+import javax.ejb.Stateful;
+import javax.enterprise.context.ApplicationScoped;
 import javax.interceptor.AroundInvoke;
+import javax.interceptor.Interceptors;
 import javax.interceptor.InvocationContext;
 
-public class LegacyInterceptor implements Serializable {
+@Stateful
+@ApplicationScoped
+@CdiIntercepted
+@Interceptors(LegacyInterceptor.class)
+public class InterceptedBean implements Serializable {
 
-    private static final long serialVersionUID = -3142706070329564629L;
+    private static final long serialVersionUID = -4444919869290540443L;
+
+    public void ping(List<String> list) {
+        list.add("InterceptedBean");
+    }
 
     @AroundInvoke
     Object aroundInvoke(InvocationContext ctx) throws Exception {
         Object[] parameters = ctx.getParameters();
         @SuppressWarnings("unchecked")
         List<String> sequence = (List<String>) parameters[0];
-        sequence.add("LegacyInterceptor");
+        sequence.add("TargetClassInterceptor");
         return ctx.proceed();
     }
 }
