+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/weld/ejb/interceptor/ordering/LegacyInterceptor.java
@@ -14,27 +14,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.as.test.integration.weld.interceptor.ordering;
+package org.jboss.as.test.integration.weld.ejb.interceptor.ordering;
 
 import java.io.Serializable;
 import java.util.List;
 
 import javax.interceptor.AroundInvoke;
-import javax.interceptor.Interceptor;
 import javax.interceptor.InvocationContext;
 
-@Interceptor
-@CdiIntercepted
-public class CdiInterceptor implements Serializable {
+public class LegacyInterceptor implements Serializable {
 
-    private static final long serialVersionUID = -5949866804898740300L;
+    private static final long serialVersionUID = -3142706070329564629L;
 
     @AroundInvoke
     Object aroundInvoke(InvocationContext ctx) throws Exception {
         Object[] parameters = ctx.getParameters();
         @SuppressWarnings("unchecked")
         List<String> sequence = (List<String>) parameters[0];
-        sequence.add("CdiInterceptor");
+        sequence.add("LegacyInterceptor");
         return ctx.proceed();
     }
 }
